refactor(anthos.dashboard): code review
- Move constants to their own module
- Avoid param reassign in software-update
