+++ packages/manager/modules/anthos/src/dashboard/access-restriction/table/constants.js
@@ -1,0 +1,5 @@
+export const TRACKING_CHUNK = 'access';
+
+export default {
+  TRACKING_CHUNK,
+};
