+++ packages/manager/modules/anthos/src/dashboard/components/order-public-ip/constants.js
@@ -4,6 +4,7 @@ export const PRICING_DURATION = 'P1M';
 export const PRICING_INTERVAL = 1;
 export const PRICING_INTERVAL_UNIT = 'month';
 export const PRODUCT_ID = 'anthos';
+export const TRACKING_CHUNK = 'order-public-ip';
 
 export default {
   ADDON_PLAN_CODE,
@@ -12,4 +13,5 @@ export default {
   PRICING_INTERVAL_UNIT,
   PRICING_INTERVAL,
   PRODUCT_ID,
+  TRACKING_CHUNK,
 };
