+++ packages/manager/modules/anthos/src/dashboard/components/order-public-ip/order-public-ip.constants.js
@@ -1,0 +1,5 @@
+export const TRACKING_CHUNK = 'order-public-ip';
+
+export default {
+  TRACKING_CHUNK,
+};
